// training_frontend/src/api/models/ExerciseLog.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Exercise } from './Exercise';
import {
    ExerciseFromJSON,
    ExerciseFromJSONTyped,
    ExerciseToJSON,
    ExerciseToJSONTyped,
} from './Exercise';

/**
 * 
 * @export
 * @interface ExerciseLog
 */
export interface ExerciseLog {
    /**
     * 
     * @type {string}
     * @memberof ExerciseLog
     */
    id: string;
    /**
     * 
     * @type {Exercise}
     * @memberof ExerciseLog
     */
    exercise: Exercise;
    /**
     * 
     * @type {number}
     * @memberof ExerciseLog
     */
    reps: number;
    /**
     * 
     * @type {number}
     * @memberof ExerciseLog
     */
    weight: number;
    /**
     * 
     * @type {number}
     * @memberof ExerciseLog
     */
    startedAt: number;
    /**
     * 
     * @type {number}
     * @memberof ExerciseLog
     */
    endedAt: number;
}

/**
 * Check if a given object implements the ExerciseLog interface.
 */
export function instanceOfExerciseLog(value: object): value is ExerciseLog {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('exercise' in value) || value['exercise'] === undefined) return false;
    if (!('reps' in value) || value['reps'] === undefined) return false;
    if (!('weight' in value) || value['weight'] === undefined) return false;
    if (!('startedAt' in value) || value['startedAt'] === undefined) return false;
    if (!('endedAt' in value) || value['endedAt'] === undefined) return false;
    return true;
}

export function ExerciseLogFromJSON(json: any): ExerciseLog {
    return ExerciseLogFromJSONTyped(json, false);
}

export function ExerciseLogFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExerciseLog {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'exercise': ExerciseFromJSON(json['exercise']),
        'reps': json['reps'],
        'weight': json['weight'],
        'startedAt': json['startedAt'],
        'endedAt': json['endedAt'],
    };
}

export function ExerciseLogToJSON(json: any): ExerciseLog {
    return ExerciseLogToJSONTyped(json, false);
}

export function ExerciseLogToJSONTyped(value?: ExerciseLog | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'exercise': ExerciseToJSON(value['exercise']),
        'reps': value['reps'],
        'weight': value['weight'],
        'startedAt': value['startedAt'],
        'endedAt': value['endedAt'],
    };
}