// training_frontend/src/api/models/PostWorkoutLogRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkoutRef } from './WorkoutRef';
import {
    WorkoutRefFromJSON,
    WorkoutRefFromJSONTyped,
    WorkoutRefToJSON,
    WorkoutRefToJSONTyped,
} from './WorkoutRef';
import type { MuscleGroup } from './MuscleGroup';
import {
    MuscleGroupFromJSON,
    MuscleGroupFromJSONTyped,
    MuscleGroupToJSON,
    MuscleGroupToJSONTyped,
} from './MuscleGroup';
import type { ExerciseLog } from './ExerciseLog';
import {
    ExerciseLogFromJSON,
    ExerciseLogFromJSONTyped,
    ExerciseLogToJSON,
    ExerciseLogToJSONTyped,
} from './ExerciseLog';

/**
 * 
 * @export
 * @interface PostWorkoutLogRequest
 */
export interface PostWorkoutLogRequest {
    /**
     * 
     * @type {string}
     * @memberof PostWorkoutLogRequest
     */
    name: string;
    /**
     * 
     * @type {WorkoutRef}
     * @memberof PostWorkoutLogRequest
     */
    baseWorkoutRef: WorkoutRef;
    /**
     * 
     * @type {number}
     * @memberof PostWorkoutLogRequest
     */
    startedAt: number;
    /**
     * 
     * @type {number}
     * @memberof PostWorkoutLogRequest
     */
    endedAt: number;
    /**
     * 
     * @type {Array<MuscleGroup>}
     * @memberof PostWorkoutLogRequest
     */
    muscleGroups: Array<MuscleGroup>;
    /**
     * 
     * @type {Array<ExerciseLog>}
     * @memberof PostWorkoutLogRequest
     */
    exercises: Array<ExerciseLog>;
}

/**
 * Check if a given object implements the PostWorkoutLogRequest interface.
 */
export function instanceOfPostWorkoutLogRequest(value: object): value is PostWorkoutLogRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('baseWorkoutRef' in value) || value['baseWorkoutRef'] === undefined) return false;
    if (!('startedAt' in value) || value['startedAt'] === undefined) return false;
    if (!('endedAt' in value) || value['endedAt'] === undefined) return false;
    if (!('muscleGroups' in value) || value['muscleGroups'] === undefined) return false;
    if (!('exercises' in value) || value['exercises'] === undefined) return false;
    return true;
}

export function PostWorkoutLogRequestFromJSON(json: any): PostWorkoutLogRequest {
    return PostWorkoutLogRequestFromJSONTyped(json, false);
}

export function PostWorkoutLogRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostWorkoutLogRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'baseWorkoutRef': WorkoutRefFromJSON(json['baseWorkoutRef']),
        'startedAt': json['startedAt'],
        'endedAt': json['endedAt'],
        'muscleGroups': ((json['muscleGroups'] as Array<any>).map(MuscleGroupFromJSON)),
        'exercises': ((json['exercises'] as Array<any>).map(ExerciseLogFromJSON)),
    };
}

export function PostWorkoutLogRequestToJSON(json: any): PostWorkoutLogRequest {
    return PostWorkoutLogRequestToJSONTyped(json, false);
}

export function PostWorkoutLogRequestToJSONTyped(value?: PostWorkoutLogRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'baseWorkoutRef': WorkoutRefToJSON(value['baseWorkoutRef']),
        'startedAt': value['startedAt'],
        'endedAt': value['endedAt'],
        'muscleGroups': ((value['muscleGroups'] as Array<any>).map(MuscleGroupToJSON)),
        'exercises': ((value['exercises'] as Array<any>).map(ExerciseLogToJSON)),
    };
}