// training_frontend/src/api/models/WorkoutDescription.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MuscleGroup } from './MuscleGroup';
import {
    MuscleGroupFromJSON,
    MuscleGroupFromJSONTyped,
    MuscleGroupToJSON,
    MuscleGroupToJSONTyped,
} from './MuscleGroup';

/**
 * 
 * @export
 * @interface WorkoutDescription
 */
export interface WorkoutDescription {
    /**
     * 
     * @type {string}
     * @memberof WorkoutDescription
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof WorkoutDescription
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof WorkoutDescription
     */
    editedAt: number;
    /**
     * 
     * @type {Array<MuscleGroup>}
     * @memberof WorkoutDescription
     */
    muscleGroups: Array<MuscleGroup>;
}

/**
 * Check if a given object implements the WorkoutDescription interface.
 */
export function instanceOfWorkoutDescription(value: object): value is WorkoutDescription {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('editedAt' in value) || value['editedAt'] === undefined) return false;
    if (!('muscleGroups' in value) || value['muscleGroups'] === undefined) return false;
    return true;
}

export function WorkoutDescriptionFromJSON(json: any): WorkoutDescription {
    return WorkoutDescriptionFromJSONTyped(json, false);
}

export function WorkoutDescriptionFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkoutDescription {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'editedAt': json['editedAt'],
        'muscleGroups': ((json['muscleGroups'] as Array<any>).map(MuscleGroupFromJSON)),
    };
}

export function WorkoutDescriptionToJSON(json: any): WorkoutDescription {
    return WorkoutDescriptionToJSONTyped(json, false);
}

export function WorkoutDescriptionToJSONTyped(value?: WorkoutDescription | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'editedAt': value['editedAt'],
        'muscleGroups': ((value['muscleGroups'] as Array<any>).map(MuscleGroupToJSON)),
    };
}